flood_alert_ml/requirements.txt
PyQt6>=6.5
requests>=2.31
scikit-learn>=1.3
numpy>=1.23
openpyxl>=3.1
python-dotenv>=1.0
pydantic>=2.0
pyqtgraph>=0.13
winrt>=1.0.0; platform_system == "Windows" and python_version < "3.13"
firebase-admin>=6.5
